requirements.txt
streamlit==1.25.0
pandas==1.5.3
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
openai==0.28.0
xlsxwriter==3.1.4
